Fixed bug in MobileBingRewardsBot.finish()
In Mobile bot's execute method, we open and close a browser for each
BingRewards account. At the end, when we call finish() to clean up any
remaining resources, we were previously trying to close a browser that had
been already closed by the execute() method. This has been rectified now.